API: value_counts consistently return int64 dtype

+++ pandas/core/arrays/string_.py
@@ -470,7 +470,7 @@ class StringArray(BaseStringArray, PandasArray):
     def value_counts(self, dropna: bool = True):
         from pandas import value_counts
 
-        return value_counts(self._ndarray, dropna=dropna).astype("Int64")
+        return value_counts(self._ndarray, dropna=dropna)
 
     def memory_usage(self, deep: bool = False) -> int:
         result = self._ndarray.nbytes
