+++ pandas/tests/arrays/integer/test_function.py
@@ -108,11 +108,11 @@ def test_stat_method(pandasmethname, kwargs):
 def test_value_counts_na():
     arr = pd.array([1, 2, 1, pd.NA], dtype="Int64")
     result = arr.value_counts(dropna=False)
-    expected = pd.Series([2, 1, 1], index=[1, 2, pd.NA], dtype="Int64")
+    expected = pd.Series([2, 1, 1], index=[1, 2, pd.NA])
     tm.assert_series_equal(result, expected)
 
     result = arr.value_counts(dropna=True)
-    expected = pd.Series([2, 1], index=[1, 2], dtype="Int64")
+    expected = pd.Series([2, 1], index=[1, 2])
     tm.assert_series_equal(result, expected)
 
 
@@ -122,7 +122,7 @@ def test_value_counts_empty():
     result = ser.value_counts()
     # TODO: The dtype of the index seems wrong (it's int64 for non-empty)
     idx = pd.Index([], dtype="object")
-    expected = pd.Series([], index=idx, dtype="Int64")
+    expected = pd.Series([], index=idx, dtype="int64")
     tm.assert_series_equal(result, expected)
 
 
@@ -130,7 +130,7 @@ def test_value_counts_with_normalize():
     # GH 33172
     ser = pd.Series([1, 2, 1, pd.NA], dtype="Int64")
     result = ser.value_counts(normalize=True)
-    expected = pd.Series([2, 1], index=[1, 2], dtype="Float64") / 3
+    expected = pd.Series([2, 1], index=[1, 2]) / 3
     tm.assert_series_equal(result, expected)
